listings and readings table migration fixed

+++ database/migrations/2021_12_20_121237_create_listings_table.php
@@ -15,7 +15,7 @@ class CreateListingsTable extends Migration
     {
         Schema::create('listings', function (Blueprint $table) {
             $table->id();
-            $table->integer("no");
+            $table->bigInteger("no");
             $table->integer("time_start");
             $table->integer("time_finish");
             $table->integer("news_no");
