+++ database/migrations/2021_12_20_121242_create_readings_table.php
@@ -15,7 +15,7 @@ class CreateReadingsTable extends Migration
     {
         Schema::create('readings', function (Blueprint $table) {
             $table->id();
-            $table->integer("no");
+            $table->bigInteger("no");
             $table->integer("time_start");
             $table->integer("time_finish");
             $table->integer("news_no");
